Resolve comments and issues

+++ src/database/migrations/2020_12_08_160641_update_favorites_table.php
@@ -27,8 +27,10 @@ class UpdateFavoritesTable extends Migration
      */
     public function down()
     {
-        Schema::table('favorites', function (Blueprint $table) {
-            $table->dropColumn(['author', 'posting_date']);
-        });
+        if (Schema::hasColumn('favorites', 'author')) {
+            Schema::table('favorites', function (Blueprint $table) {
+                $table->dropColumn(['author', 'posting_date']);
+            });
+        }
     }
 }
